go-github: move event types into event_types.go
fixes #283
